test: add raw TCP integration test type
Some server behavior can't be exercised through a Redis client (ioredis /
node-redis): inline commands, malformed frames, exact response bytes, and
explicit request pipelining. Clients only ever speak well-formed RESP and
hide the wire bytes, so these paths previously had no integration coverage.

Add a raw-TCP integration harness under tests-integration/raw-tcp/:

- raw-connection.ts: a RawRedisConnection over a bare socket plus a RESP2/3
  reply parser. This parser was not newly written — it was extracted from the
  inline copy in connection-integration.test.ts and promoted to a shared
  module so both suites reuse it. connection-integration now imports it.
- protocol.test.ts: wire-level tests (byte-exact PING, inline commands,
  quoted inline args, single-write pipelining, null bulk) that run against
  both backends, since the in-process mock server and a real redis-server
  both listen on an ordinary TCP port.
- TestRunner.setupRawStandalone(): hands back a standalone port (mock: in-
  process Resp2Server; real: REDIS_STANDALONE_PORT or a spawned child) with
  no client attached, so a test can open its own socket.
- package.json: test:integration:raw:{mock,real} scripts for targeted runs;
  the raw-tcp dir is already covered by the existing **-glob integration runs.

Connections are tracked and closed before cleanup() because the mock's
net.Server.close() only resolves once every client socket has ended.
