+++ package.json
@@ -49,6 +49,8 @@
     "test": "tsc --noEmit && glob -c \"node --enable-source-maps --import tsx --no-warnings --test\" \"./tests/**/*.test.ts\"",
     "test:integration:mock": "tsc --noEmit && TEST_BACKEND=mock glob -c \"node --enable-source-maps --import tsx --no-warnings --test\" \"./tests-integration/**/*.test.ts\"",
     "test:integration:real": "tsc --noEmit && npm run clean:redis && TEST_BACKEND=real glob -c \"node --enable-source-maps --import tsx --no-warnings --test-concurrency 1 --test-timeout 30000 --test\" \"./tests-integration/**/*.test.ts\"",
+    "test:integration:raw:mock": "tsc --noEmit && TEST_BACKEND=mock glob -c \"node --enable-source-maps --import tsx --no-warnings --test\" \"./tests-integration/raw-tcp/**/*.test.ts\"",
+    "test:integration:raw:real": "tsc --noEmit && TEST_BACKEND=real glob -c \"node --enable-source-maps --import tsx --no-warnings --test-concurrency 1 --test-timeout 30000 --test\" \"./tests-integration/raw-tcp/**/*.test.ts\"",
     "test:all": "npm run test && npm run test:integration:mock && npm run test:integration:real",
     "clean:redis": "redis-cli --cluster call 127.0.0.1:30000 FLUSHALL || redis-cli FLUSHALL || echo 'Redis cleanup attempted'",
     "lint": "eslint .",
